+++ yandex/data_source_yandex_alb_backend_group.go
@@ -275,6 +275,11 @@ func dataSourceLoadBalancingConfig() *schema.Schema {
 					Optional: true,
 					Computed: true,
 				},
+				"mode": {
+					Type:     schema.TypeString,
+					Optional: true,
+					Computed: true,
+				},
 			},
 		},
 	}