+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.76.0 (Unreleased)
 * alb: `send` and `receive` attributes can be empty now in `yandex_alb_backend_group` resource and data source
 * alb: fixed `yandex_alb_load_balancer` resource and data source documentation
+* alb: support `mode` attribute in `load_balancing_config` entity in `yandex_alb_backend_group` resource and data source
 
 FEATURES:
 * compute: support for creating `yandex_compute_instance` with local disks